requirements.txt
fastapi
pandas
playsound
request
uvicorn